dealing with C++ warnings
git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@2734 3d70eeeb-363e-0410-a505-8a46323a89f2

+++ src/hest/test/bday.c
@@ -81,7 +81,7 @@ main(int argc, char *argv[]) {
   crct = 1-crct;
   yes = 0;
   E = 1;
-  airSrand48(airTime());
+  airSrand48((int)airTime());
   while (E) {
     yes += runexp(year, N, M);
     if (!(E % P)) {
